+++ consensus/hotstuff_leader.go
@@ -2,6 +2,7 @@ package consensus
 
 import (
 	"errors"
+	"fmt"
 
 	consensusTelemetry "github.com/pokt-network/pocket/consensus/telemetry"
 	typesCons "github.com/pokt-network/pocket/consensus/types"
@@ -9,8 +10,6 @@ import (
 	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 	"github.com/pokt-network/pocket/shared/modules"
 )
-
-// CONSOLIDATE: Last/Prev & AppHash/StateHash/BlockHash
 
 type HotstuffLeaderMessageHandler struct{}
 
@@ -43,7 +42,7 @@ func (handler *HotstuffLeaderMessageHandler) HandleNewRoundMessage(m *consensusM
 	// DISCUSS: Do we need to pause for `MinBlockFreqMSec` here to let more transactions or should we stick with optimistic responsiveness?
 
 	if err := m.didReceiveEnoughMessageForStep(NewRound); err != nil {
-		m.logger.Info().Msg(typesCons.OptimisticVoteCountWaiting(NewRound, err.Error()))
+		m.logger.Info().Fields(msgToLoggingFields(msg)).Msgf("⏳ Waiting ⏳for more messages; %s", err.Error())
 		return
 	}
 
@@ -130,7 +129,7 @@ func (handler *HotstuffLeaderMessageHandler) HandlePrepareMessage(m *consensusMo
 	m.logger.Debug().Fields(m.hotstuffMsgLogHelper(msg)).Msg("HandlePrepare Leader checking enough votes")
 
 	if err := m.didReceiveEnoughMessageForStep(Prepare); err != nil {
-		m.logger.Info().Msg(typesCons.OptimisticVoteCountWaiting(Prepare, err.Error()))
+		m.logger.Info().Fields(msgToLoggingFields(msg)).Msgf("⏳ Waiting ⏳for more messages; %s", err.Error())
 		return
 	}
 
@@ -187,7 +186,7 @@ func (handler *HotstuffLeaderMessageHandler) HandlePrecommitMessage(m *consensus
 	m.logger.Debug().Fields(m.hotstuffMsgLogHelper(msg)).Msg("HandlePreCommit Leader checking enough votes")
 
 	if err := m.didReceiveEnoughMessageForStep(PreCommit); err != nil {
-		m.logger.Info().Msg(typesCons.OptimisticVoteCountWaiting(PreCommit, err.Error()))
+		m.logger.Info().Fields(msgToLoggingFields(msg)).Msgf("⏳ Waiting ⏳for more messages; %s", err.Error())
 		return
 	}
 
@@ -246,7 +245,7 @@ func (handler *HotstuffLeaderMessageHandler) HandleCommitMessage(m *consensusMod
 	m.logger.Debug().Fields(m.hotstuffMsgLogHelper(msg)).Msg("HandleCommit Leader checking enough votes")
 
 	if err := m.didReceiveEnoughMessageForStep(Commit); err != nil {
-		m.logger.Info().Msg(typesCons.OptimisticVoteCountWaiting(Commit, err.Error()))
+		m.logger.Info().Fields(msgToLoggingFields(msg)).Msgf("⏳ Waiting ⏳for more messages; %s", err.Error())
 		return
 	}
 
@@ -423,15 +422,21 @@ func (m *consensusModule) prepareBlock(qc *typesCons.QuorumCertificate) (*coreTy
 		return nil, typesCons.ErrReplicaPrepareBlock
 	}
 
+	utilityUnitOfWork := m.utilityUnitOfWork
+	if utilityUnitOfWork == nil {
+		return nil, fmt.Errorf("utility unit of work is nil")
+	}
+
 	maxTxBytes := m.consCfg.MaxMempoolBytes
 
-	leaderUOW, ok := m.utilityUnitOfWork.(modules.LeaderUtilityUnitOfWork)
+	leaderUOW, ok := utilityUnitOfWork.(modules.LeaderUtilityUnitOfWork)
 	if !ok {
 		return nil, errors.New("invalid utility unitOfWork, should be of type LeaderUtilityUnitOfWork")
 	}
 
 	// Reap the mempool for transactions to be applied in this block
 	stateHash, txs, err := leaderUOW.CreateAndApplyProposalBlock(m.privateKey.Address(), maxTxBytes)
+
 	if err != nil {
 		return nil, err
 	}
@@ -441,6 +446,7 @@ func (m *consensusModule) prepareBlock(qc *typesCons.QuorumCertificate) (*coreTy
 	if err != nil {
 		return nil, err
 	}
+	defer readCtx.Release()
 
 	prevBlockHash, err := readCtx.GetBlockHash(prevHeight)
 	if err != nil {
@@ -468,7 +474,7 @@ func (m *consensusModule) prepareBlock(qc *typesCons.QuorumCertificate) (*coreTy
 	m.logger.Debug().Msgf("Prepared a new block, QC is: %c, len: %d", qcBytes, len(qcBytes))
 
 	// Set the proposal block in the persistence context
-	if err := m.utilityUnitOfWork.SetProposalBlock(blockHeader.StateHash, blockHeader.ProposerAddress, block.Transactions); err != nil {
+	if err := utilityUnitOfWork.SetProposalBlock(blockHeader.StateHash, blockHeader.ProposerAddress, block.Transactions); err != nil {
 		return nil, err
 	}