+++ consensus/hotstuff_replica.go
@@ -10,8 +10,6 @@ import (
 	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 )
 
-// CONSOLIDATE: Terminology of `appHash` and `stateHash`
-
 type HotstuffReplicaMessageHandler struct{}
 
 var (
@@ -275,6 +273,11 @@ func (m *consensusModule) validateProposal(msg *typesCons.HotstuffMessage) error
 // This helper applies the block metadata to the utility & persistence layers
 func (m *consensusModule) applyBlock(block *coreTypes.Block) error {
 	blockHeader := block.BlockHeader
+	utilityUnitOfWork := m.utilityUnitOfWork
+	if utilityUnitOfWork == nil {
+		return fmt.Errorf("utility unit of work is nil")
+	}
+
 	// Set the proposal block in the persistence context
 
 	fmt.Printf("Apply block called with block transactions: %x \n", block.Transactions)
@@ -291,12 +294,15 @@ func (m *consensusModule) applyBlock(block *coreTypes.Block) error {
 		fmt.Println("block.Transactions is nil")
 	}
 
-	if err := m.utilityUnitOfWork.SetProposalBlock(blockHeader.StateHash, blockHeader.ProposerAddress, block.Transactions); err != nil {
+	//if err := m.utilityUnitOfWork.SetProposalBlock(blockHeader.StateHash, blockHeader.ProposerAddress, block.Transactions); err != nil {
+
+	if err := utilityUnitOfWork.SetProposalBlock(blockHeader.StateHash, blockHeader.ProposerAddress, block.Transactions); err != nil {
+
 		return err
 	}
 
 	// Apply all the transactions in the block and get the stateHash
-	stateHash, _, err := m.utilityUnitOfWork.ApplyBlock()
+	stateHash, _, err := utilityUnitOfWork.ApplyBlock()
 	if err != nil {
 		return err
 	}
@@ -305,7 +311,7 @@ func (m *consensusModule) applyBlock(block *coreTypes.Block) error {
 	m.logger.Info().Msg("finishing applyBlock")
 
 	if blockHeader.StateHash != stateHash {
-		return typesCons.ErrInvalidAppHash(blockHeader.StateHash, stateHash)
+		return typesCons.ErrInvalidStateHash(blockHeader.StateHash, stateHash)
 	}
 
 	return nil
@@ -356,7 +362,7 @@ func (m *consensusModule) validateQuorumCertificate(qc *typesCons.QuorumCertific
 		}
 		numValid++
 	}
-	if err := m.isOptimisticThresholdMet(numValid, validators); err != nil {
+	if err := m.validateOptimisticThresholdMet(numValid, validators); err != nil {
 		return err
 	}