+++ persistence/genesis.go
@@ -28,23 +28,24 @@ func (m *persistenceModule) populateGenesisState(state *genesis.GenesisState) {
 		return nil
 	}
 
-	rwContext, err := m.NewRWContext(0)
+	rwCtx, err := m.NewRWContext(0)
 	if err != nil {
 		m.logger.Fatal().Err(err).Msg("an error occurred creating the rwContext for the genesis state")
 	}
+	// NB: Not calling `defer rwCtx.Release()` because we `Commit`, which releases the tx below
 
 	for _, acc := range state.GetAccounts() {
 		addrBz, err := hex.DecodeString(acc.GetAddress())
 		if err != nil {
 			m.logger.Fatal().Err(err).Str("address", acc.GetAddress()).Msg("an error occurred converting address to bytes")
 		}
-		err = rwContext.SetAccountAmount(addrBz, acc.GetAmount())
+		err = rwCtx.SetAccountAmount(addrBz, acc.GetAmount())
 		if err != nil {
 			m.logger.Fatal().Err(err).Str("address", acc.GetAddress()).Msg("an error occurred inserting an acc in the genesis state")
 		}
 	}
 	for _, pool := range state.GetPools() {
-		err = rwContext.InsertPool(pool.GetAddress(), pool.GetAmount()) // pool.GetAddress() returns the pool's semantic name
+		err = rwCtx.InsertPool(pool.GetAddress(), pool.GetAmount()) // pool.GetAddress() returns the pool's semantic name
 		if err != nil {
 			m.logger.Fatal().Err(err).Str("address", pool.GetAddress()).Msg("an error occurred inserting an pool in the genesis state")
 		}
@@ -60,27 +61,27 @@ func (m *persistenceModule) populateGenesisState(state *genesis.GenesisState) {
 			Name:   "app",
 			Getter: state.GetApplications,
 			InsertFn: func(address, publicKey, output []byte, paused bool, status int32, serviceURL, stakedTokens string, chains []string, pausedHeight, unstakingHeight int64) error {
-				return rwContext.InsertApp(address, publicKey, output, paused, status, stakedTokens, chains, pausedHeight, unstakingHeight)
+				return rwCtx.InsertApp(address, publicKey, output, paused, status, stakedTokens, chains, pausedHeight, unstakingHeight)
 			},
 			Pool: coreTypes.Pools_POOLS_APP_STAKE,
 		},
 		{
 			Name:     "servicer",
 			Getter:   state.GetServicers,
-			InsertFn: rwContext.InsertServicer,
+			InsertFn: rwCtx.InsertServicer,
 			Pool:     coreTypes.Pools_POOLS_SERVICER_STAKE,
 		},
 		{
 			Name:     "fisherman",
 			Getter:   state.GetFishermen,
-			InsertFn: rwContext.InsertFisherman,
+			InsertFn: rwCtx.InsertFisherman,
 			Pool:     coreTypes.Pools_POOLS_FISHERMAN_STAKE,
 		},
 		{
 			Name:   "validator",
 			Getter: state.GetValidators,
 			InsertFn: func(address, publicKey, output []byte, paused bool, status int32, serviceURL, stakedTokens string, chains []string, pausedHeight, unstakingHeight int64) error {
-				return rwContext.InsertValidator(address, publicKey, output, paused, status, serviceURL, stakedTokens, pausedHeight, unstakingHeight)
+				return rwCtx.InsertValidator(address, publicKey, output, paused, status, serviceURL, stakedTokens, pausedHeight, unstakingHeight)
 			},
 			Pool: coreTypes.Pools_POOLS_VALIDATOR_STAKE,
 		},
@@ -110,16 +111,16 @@ func (m *persistenceModule) populateGenesisState(state *genesis.GenesisState) {
 		}
 	}
 
-	if err = rwContext.InitGenesisParams(state.Params); err != nil {
+	if err = rwCtx.InitGenesisParams(state.Params); err != nil {
 		log.Fatalf("an error occurred initializing params: %s", err.Error())
 	}
 
-	if err = rwContext.InitFlags(); err != nil { // TODO (Team) use flags from genesis file not hardcoded
+	if err = rwCtx.InitFlags(); err != nil { // TODO (Team) use flags from genesis file not hardcoded
 		m.logger.Fatal().Err(err).Msg("an error occurred initializing flags")
 	}
 
 	// Updates all the merkle trees
-	stateHash, err := rwContext.ComputeStateHash()
+	stateHash, err := rwCtx.ComputeStateHash()
 	if err != nil {
 		m.logger.Fatal().Err(err).Msg("an error occurred updating the app hash during genesis")
 	}
@@ -127,7 +128,7 @@ func (m *persistenceModule) populateGenesisState(state *genesis.GenesisState) {
 
 	// This updates the DB, blockstore, and commits the genesis state.
 	// Note that the `quorumCert for genesis` is nil.
-	if err = rwContext.Commit(nil, nil, nil); err != nil {
+	if err = rwCtx.Commit(nil, nil, nil); err != nil {
 		m.logger.Fatal().Err(err).Msg("an error occurred committing the genesis state to the DB")
 	}
 }